Decompose TurboQuant: pull SORF into ScalarFn, replace codes/centroids with DictArray
Replace the monolithic TurboQuantArray VTable with a composable tree of
standard array types:

  ScalarFnArray(L2Denorm, [
    ScalarFnArray(SorfTransform, [FSL(Dict(codes, centroids))]),
    norms
  ])

Each layer is independently swappable:
- Normalization: L2Denorm stores per-vector norms
- Rotation: SorfTransform records the SORF rotation via a deterministic
  PRNG seed in SorfOptions; reconstructs RotationMatrix at decode time
- Quantization: DictArray(codes, centroids) wrapped in FixedSizeListArray

Decompression is automatic: executing the outer array walks the
ScalarFn tree (Dict dequantizes → SorfTransform inverse-rotates →
L2Denorm re-applies norms).

New:
- SorfTransform ScalarFnVTable in scalar_fns/sorf_transform.rs
- build_quantized_fsl() for Dict construction, separate from SORF wrapping
- SorfTransform isolation test exercising the ScalarFn wiring directly
- Invalid bit_width rejection test

Removed:
- TurboQuantArray VTable, metadata, compute pushdowns, slots, data,
  decompress.rs, serde roundtrip tests (~2,500 lines)
- TurboQuant from vortex-file allowed encodings (no longer a custom
  array encoding — uses standard ScalarFnArray/FSL/Dict)

Moved:
- centroids.rs and rotation.rs promoted from array/ subdirectory
- Tests split into tests/ directory module (roundtrip, compute,
  structural, nullable)
