+++ vortex-tensor/src/encodings/turboquant/scheme/compress.rs
@@ -15,9 +15,9 @@ use vortex_array::IntoArray;
 use vortex_array::arrays::Extension;
 use vortex_array::arrays::FixedSizeListArray;
 use vortex_array::arrays::PrimitiveArray;
+use vortex_array::arrays::dict::DictArray;
 use vortex_array::arrays::extension::ExtensionArrayExt;
 use vortex_array::arrays::fixed_size_list::FixedSizeListArrayExt;
-use vortex_array::dtype::DType;
 use vortex_array::dtype::Nullability;
 use vortex_array::dtype::PType;
 use vortex_array::validity::Validity;
@@ -26,15 +26,17 @@ use vortex_error::VortexExpect;
 use vortex_error::VortexResult;
 use vortex_error::vortex_bail;
 use vortex_error::vortex_ensure;
-use vortex_fastlanes::bitpack_compress::bitpack_encode;
 
-use crate::encodings::turboquant::TurboQuant;
-use crate::encodings::turboquant::array::centroids::compute_centroid_boundaries;
-use crate::encodings::turboquant::array::centroids::find_nearest_centroid;
-use crate::encodings::turboquant::array::centroids::get_centroids;
-use crate::encodings::turboquant::array::rotation::RotationMatrix;
-use crate::encodings::turboquant::vtable::TurboQuantArray;
+use crate::encodings::turboquant::MAX_BIT_WIDTH;
+use crate::encodings::turboquant::MIN_DIMENSION;
+use crate::encodings::turboquant::centroids::compute_centroid_boundaries;
+use crate::encodings::turboquant::centroids::find_nearest_centroid;
+use crate::encodings::turboquant::centroids::get_centroids;
+use crate::encodings::turboquant::rotation::RotationMatrix;
 use crate::scalar_fns::l2_denorm::validate_l2_normalized_rows;
+use crate::scalar_fns::sorf_transform::SorfOptions;
+use crate::scalar_fns::sorf_transform::SorfTransform;
+use crate::vector::AnyVector;
 
 /// Configuration for TurboQuant encoding.
 #[derive(Clone, Debug)]
@@ -50,7 +52,7 @@ pub struct TurboQuantConfig {
 impl Default for TurboQuantConfig {
     fn default() -> Self {
         Self {
-            bit_width: TurboQuant::MAX_BIT_WIDTH,
+            bit_width: MAX_BIT_WIDTH,
             seed: Some(42),
             num_rounds: 3,
         }
@@ -93,7 +95,6 @@ fn extract_f32_elements(
 
 /// Shared intermediate results from the quantization loop.
 struct QuantizationResult {
-    rotation: RotationMatrix,
     centroids: Vec<f32>,
     all_indices: BufferMut<u8>,
     padded_dim: usize,
@@ -145,48 +146,44 @@ fn turboquant_quantize_core(
     }
 
     Ok(QuantizationResult {
-        rotation,
         centroids,
         all_indices,
         padded_dim,
     })
 }
 
-/// Build a `TurboQuantArray` from quantization results.
+/// Build a quantized representation: `FSL(DictArray(codes, centroids), padded_dim)`.
 ///
-/// The `ext_dtype` must be a non-nullable [`Vector`](crate::vector::Vector) extension dtype.
-fn build_turboquant(
+/// This is a Dict-encoded FixedSizeList where each row of `padded_dim` u8 codes
+/// indexes into the centroid codebook. The Dict can be independently sliced, taken,
+/// or executed (dequantized) without knowledge of the rotation.
+fn build_quantized_fsl(
     num_rows: usize,
-    core: QuantizationResult,
-    ext_dtype: &DType,
-) -> VortexResult<TurboQuantArray> {
-    let padded_dim = core.padded_dim;
+    all_indices: BufferMut<u8>,
+    centroids: &[f32],
+    padded_dim: usize,
+) -> VortexResult<ArrayRef> {
+    let codes = PrimitiveArray::new::<u8>(all_indices.freeze(), Validity::NonNullable);
+
+    let mut centroids_buf = BufferMut::<f32>::with_capacity(centroids.len());
+    centroids_buf.extend_from_slice(centroids);
+    let centroids_array = PrimitiveArray::new::<f32>(centroids_buf.freeze(), Validity::NonNullable);
+
+    let dict = DictArray::try_new(codes.into_array(), centroids_array.into_array())?;
+
     let padded_dim_u32 =
         u32::try_from(padded_dim).vortex_expect("padded_dim stays representable as u32");
-    let codes_elements =
-        PrimitiveArray::new::<u8>(core.all_indices.freeze(), Validity::NonNullable);
-    let codes = FixedSizeListArray::try_new(
-        codes_elements.into_array(),
+    Ok(FixedSizeListArray::try_new(
+        dict.into_array(),
         padded_dim_u32,
         Validity::NonNullable,
         num_rows,
     )?
-    .into_array();
-
-    // TODO(perf): `get_centroids` returns Vec<f32>; could avoid the copy by
-    // supporting Buffer::from(Vec<T>) or caching as Buffer directly.
-    let mut centroids_buf = BufferMut::<f32>::with_capacity(core.centroids.len());
-    centroids_buf.extend_from_slice(&core.centroids);
-    let centroids_array =
-        PrimitiveArray::new::<f32>(centroids_buf.freeze(), Validity::NonNullable).into_array();
-
-    let rotation_signs = bitpack_rotation_signs(&core.rotation)?;
-
-    TurboQuant::try_new_array(ext_dtype.clone(), codes, centroids_array, rotation_signs)
+    .into_array())
 }
 
 /// Encode a non-nullable, L2-normalized [`Vector`](crate::vector::Vector) extension array into a
-/// [`TurboQuantArray`].
+/// `ScalarFnArray(SorfTransform, [FSL(Dict(codes, centroids))])`.
 ///
 /// The input must be a non-nullable Vector extension array whose rows are already unit-norm.
 /// **Null vectors are not supported.** The caller must normalize and strip nullability before
@@ -196,9 +193,9 @@ fn build_turboquant(
 /// [`turboquant_encode_unchecked`] to skip this check when the caller has just performed
 /// normalization.
 ///
-/// The returned array is a plain [`TurboQuantArray`] that decompresses to unit-norm vectors.
-/// The caller is responsible for wrapping it in an [`L2Denorm`] ScalarFnArray if the original
-/// magnitudes need to be restored.
+/// The returned array is a `SorfTransform` ScalarFnArray wrapping `FSL(Dict)` that decompresses
+/// to unit-norm vectors. The caller is responsible for wrapping it in an [`L2Denorm`] ScalarFnArray
+/// if the original magnitudes need to be restored.
 ///
 /// [`normalize_as_l2_denorm`]: crate::scalar_fns::l2_denorm::normalize_as_l2_denorm
 /// [`L2Denorm`]: crate::scalar_fns::l2_denorm::L2Denorm
@@ -221,7 +218,8 @@ pub fn turboquant_encode(
 }
 
 /// Encode a non-nullable, L2-normalized [`Vector`](crate::vector::Vector) extension array into a
-/// [`TurboQuantArray`], without validating the unit-norm precondition.
+/// `ScalarFnArray(SorfTransform, [FSL(Dict(codes, centroids))])`, without validating the unit-norm
+/// precondition.
 ///
 /// # Safety
 ///
@@ -242,71 +240,55 @@ pub unsafe fn turboquant_encode_unchecked(
     let fsl = storage.clone().execute::<FixedSizeListArray>(ctx)?;
 
     vortex_ensure!(
-        config.bit_width >= 1 && config.bit_width <= TurboQuant::MAX_BIT_WIDTH,
-        "bit_width must be 1-{}, got {}",
-        TurboQuant::MAX_BIT_WIDTH,
+        config.bit_width >= 1 && config.bit_width <= MAX_BIT_WIDTH,
+        "bit_width must be 1-{MAX_BIT_WIDTH}, got {}",
         config.bit_width
     );
     let dimension = fsl.list_size();
     vortex_ensure!(
-        dimension >= TurboQuant::MIN_DIMENSION,
-        "TurboQuant requires dimension >= {}, got {dimension}",
-        TurboQuant::MIN_DIMENSION
+        dimension >= MIN_DIMENSION,
+        "TurboQuant requires dimension >= {MIN_DIMENSION}, got {dimension}",
     );
 
-    if fsl.is_empty() {
-        let padded_dim = dimension.next_power_of_two();
-        let empty_codes = FixedSizeListArray::try_new(
-            PrimitiveArray::empty::<u8>(Nullability::NonNullable).into_array(),
-            padded_dim,
-            Validity::NonNullable,
-            0,
-        )?;
-
-        let empty_centroids = PrimitiveArray::empty::<f32>(Nullability::NonNullable);
-        let empty_signs = FixedSizeListArray::try_new(
-            PrimitiveArray::empty::<u8>(Nullability::NonNullable).into_array(),
-            padded_dim,
-            Validity::NonNullable,
-            0,
-        )?;
-
-        return Ok(TurboQuant::try_new_array(
-            ext_dtype,
-            empty_codes.into_array(),
-            empty_centroids.into_array(),
-            empty_signs.into_array(),
-        )?
-        .into_array());
-    }
+    let vector_metadata = ext_dtype.as_extension().metadata::<AnyVector>();
+    let element_ptype = vector_metadata.element_ptype();
 
     let seed = config.seed.unwrap_or(42);
     let num_rows = fsl.len();
+
+    if fsl.is_empty() {
+        let padded_dim = dimension.next_power_of_two();
+        let empty_codes = PrimitiveArray::empty::<u8>(Nullability::NonNullable);
+        let empty_centroids = PrimitiveArray::empty::<f32>(Nullability::NonNullable);
+        let empty_dict =
+            DictArray::try_new(empty_codes.into_array(), empty_centroids.into_array())?;
+        let empty_fsl = FixedSizeListArray::try_new(
+            empty_dict.into_array(),
+            padded_dim,
+            Validity::NonNullable,
+            0,
+        )?;
+
+        let sorf_options = SorfOptions {
+            seed,
+            num_rounds: config.num_rounds,
+            dimension,
+            element_ptype,
+        };
+        return Ok(
+            SorfTransform::try_new_array(&sorf_options, empty_fsl.into_array(), 0)?.into_array(),
+        );
+    }
+
     let core = turboquant_quantize_core(&fsl, seed, config.bit_width, config.num_rounds, ctx)?;
+    let quantized_fsl =
+        build_quantized_fsl(num_rows, core.all_indices, &core.centroids, core.padded_dim)?;
 
-    Ok(build_turboquant(num_rows, core, &ext_dtype)?.into_array())
-}
-
-/// Export rotation signs as a `FixedSizeListArray` wrapping a 1-bit [`BitPackedArray`].
-///
-/// The rotation matrix's `num_rounds * padded_dim` sign values are exported as 0/1 u8 values in
-/// inverse application order, bitpacked to 1 bit per sign, then wrapped in a
-/// `FixedSizeListArray` with `list_size = padded_dim` and `len = num_rounds`.
-fn bitpack_rotation_signs(rotation: &RotationMatrix) -> VortexResult<ArrayRef> {
-    let signs_u8 = rotation.export_inverse_signs_u8();
-    let num_rounds = rotation.num_rounds();
-    let padded_dim = u32::try_from(rotation.padded_dim()).vortex_expect("padded_dim fits in u32");
-
-    let mut buf = BufferMut::<u8>::with_capacity(signs_u8.len());
-    buf.extend_from_slice(&signs_u8);
-    let prim = PrimitiveArray::new::<u8>(buf.freeze(), Validity::NonNullable);
-    let bitpacked = bitpack_encode(&prim, 1, None)?;
-
-    let fsl = FixedSizeListArray::try_new(
-        bitpacked.into_array(),
-        padded_dim,
-        Validity::NonNullable,
-        num_rounds,
-    )?;
-    Ok(fsl.into_array())
+    let sorf_options = SorfOptions {
+        seed,
+        num_rounds: config.num_rounds,
+        dimension,
+        element_ptype,
+    };
+    Ok(SorfTransform::try_new_array(&sorf_options, quantized_fsl, num_rows)?.into_array())
 }
